Move ColorRepresentation to Color.d.ts
