+++ types/three/src/Three.d.ts
@@ -3,7 +3,6 @@
  */
 export * from './constants';
 export * from './Three.Legacy';
-export * from './utils';
 /**
  * Animation
  */
