+++ types/three/src/helpers/ArrowHelper.d.ts
@@ -2,7 +2,7 @@ import { Vector3 } from './../math/Vector3';
 import { Line } from './../objects/Line';
 import { Mesh } from './../objects/Mesh';
 import { Object3D } from './../core/Object3D';
-import { ColorRepresentation } from '../utils';
+import { ColorRepresentation } from '../math/Color';
 
 // Extras / Helpers /////////////////////////////////////////////////////////////////////
