+++ types/three/src/helpers/DirectionalLightHelper.d.ts
@@ -2,7 +2,7 @@ import { DirectionalLight } from './../lights/DirectionalLight';
 import { Line } from './../objects/Line';
 import { Matrix4 } from './../math/Matrix4';
 import { Object3D } from './../core/Object3D';
-import { ColorRepresentation } from '../utils';
+import { ColorRepresentation } from '../math/Color';
 
 export class DirectionalLightHelper extends Object3D {
     /**
