+++ types/three/src/helpers/HemisphereLightHelper.d.ts
@@ -1,9 +1,8 @@
 import { HemisphereLight } from './../lights/HemisphereLight';
-import { Color } from './../math/Color';
 import { Matrix4 } from './../math/Matrix4';
 import { MeshBasicMaterial } from './../materials/MeshBasicMaterial';
 import { Object3D } from './../core/Object3D';
-import { ColorRepresentation } from '../utils';
+import { ColorRepresentation } from '../math/Color';
 
 export class HemisphereLightHelper extends Object3D {
     constructor(light: HemisphereLight, size: number, color?: ColorRepresentation);
