+++ types/three/src/lights/AmbientLight.d.ts
@@ -1,4 +1,4 @@
-import { ColorRepresentation } from '../utils';
+import { ColorRepresentation } from '../math/Color';
 import { Light } from './Light';
 
 /**
