+++ types/three/src/lights/HemisphereLight.d.ts
@@ -1,7 +1,6 @@
-import { Color } from './../math/Color';
+import { Color, ColorRepresentation } from './../math/Color';
 import { Vector3 } from '../math/Vector3';
 import { Light } from './Light';
-import { ColorRepresentation } from '../utils';
 
 export class HemisphereLight extends Light {
     /**
