+++ types/three/src/lights/HemisphereLightProbe.d.ts
@@ -1,4 +1,4 @@
-import { ColorRepresentation } from '../utils';
+import { ColorRepresentation } from '../math/Color';
 import { LightProbe } from './LightProbe';
 
 export class HemisphereLightProbe extends LightProbe {
