+++ types/three/src/materials/MeshBasicMaterial.d.ts
@@ -1,8 +1,7 @@
-import { Color } from './../math/Color';
+import { Color, ColorRepresentation } from './../math/Color';
 import { Texture } from './../textures/Texture';
 import { MaterialParameters, Material } from './Material';
 import { Combine } from '../constants';
-import { ColorRepresentation } from '../utils';
 /**
  * parameters is an object with one or more properties defining the material's appearance.
  */
